requirement.txt
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
filelock==3.8.0
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.35.0
huggingface-hub==0.8.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.5.3
numpy==1.23.2
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
regex==2022.7.25
requests==2.28.1
six==1.16.0
tokenizers==0.12.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.64.0
transformers==4.21.1
typing_extensions==4.3.0
urllib3==1.26.11
Werkzeug==2.2.2
zipp==3.8.1
coloredlogs==15.0.1
flatbuffers==2.0.7
humanfriendly==10.0
mpmath==1.2.1
protobuf==3.20.1
scipy==1.9.1
seaborn==0.12.0
sympy==1.11.1